2008-01-10 11:47 UTC+0100 Miguel Angel Marchuet <miguelangel/at/marchuet.net>   * contrib\rdd_ads\rddads.h   * include\hbapi.h   * include\hbapiitm.h   * include\hbapirdd.h   * include\hbrddbmc.h   * include\hbrddcdx.h   * include\hbrdddbf.h   * include\hbrdddel.h   * include\hbrddntx.h   * include\hbrddsdf.h   * source\rdd\bmdbfcdx\bmdbfcdx1.c   * source\rdd\dbf1.c   * source\rdd\dbfcdx\dbfcdx1.c   * source\rdd\dbffpt\dbffpt1.c   * source\compiler\harbour.slx   * source\compiler\harbour.sly   * source\rtl\valtype.c   * source\rdd\workarea.c     + Adding NULL fields capabilities in progress
  * include\usrrdd.ch
  * source\rdd\usrrdd\usrrdd.c
    + Added support to incremental fields and flags.

  * include\hbtrace.h
    ! fixed miror redefine warnings.

  * include\thread.h
    * Introduced pEnd for more quick stack control.

  * source\common\hbdate.c
    ! fixed TRACE in hb_timeDecStr( char * szTime, LONG lSeconds )

  * source\common\hbfopen.c
    + added hb_fsNameConv nedded to compile PP aplicacion.

  * source\rdd\usrrdd\rdds\adordd.prg
    ! corrected some bugs when catalog can't be created.

  * source\rdd\wafunc.c
    * changed hb_rddOpenTable to use hb_rddSelectFirstAvailable() instead of hb_rddSelectWorkAreaNumber( 0 );

  * source\vm\hvm.c
    ! fixed HB_P_POPGLOBAL missing ;
    ! fixed TRACE in hb_vmPushMacroBlock( BYTE * pCode )
    ! fixed TRACE in hb_xvmLocalSetInt( int iLocal, int iVal )
    ! fixed TRACE in hb_xvmLocalSetStr( int iLocal, const char * pVal, ULONG ulLen )
    * hb_vmPushSymbol( PHB_SYMB pSym ), hb_vmPushLocalByRef( SHORT iLocal ) to use new hb_stackAllocItem()
    * hb_vmPopLocal( SHORT iLocal )

  * include\hbstack.h
  * source\vm\arrays.c
  * source\vm\estack.c
  * source\vm\fastitem.c
  * source\vm\garbage.c
  * source\vm\hvm.c
  * source\vm\itemapi.c
    ! fixed flag type HB_IT_EXTREF, HB_IT_MEMVAR.
    * Used if else if instead of switch over type comparison, because type sometimes has flags as
      HB_IT_EXTREF, HB_IT_MEMVAR, HB_IT_BYREF or HB_IT_NULL conbined with other types.
    * Introduced pEnd for more quick stack control.
    + added  hb_stackItemBasePtr().
    * Removed some random GPF's when conbined types was used.

  * source\rdd\usrrdd\rdds\vfpcdx.prg
    ! some minor sintax fixes

+++ xharbour/source/rdd/dbfcdx/dbfcdx1.c
@@ -1,5 +1,5 @@
 /*
- * $Id: dbfcdx1.c,v 1.270 2007/10/31 08:35:10 marchuet Exp $
+ * $Id: dbfcdx1.c,v 1.271 2007/11/20 16:57:12 marchuet Exp $
  */
 
 /*
@@ -2339,7 +2339,7 @@ static void hb_cdxPageStore( LPCDXPAGE pPage )
    else if ( pPage->iKeys > pPage->TagParent->MaxKeys )
       hb_cdxErrInternal( "hb_cdxPageStore: number of keys exceed!." );
 #endif
-   HB_PUT_LE_UINT16( pPage->node.intNode.attr, pPage->PageType );
+   HB_PUT_LE_UINT16( pPage->node.intNode.attr, ( UINT16 ) pPage->PageType );
    HB_PUT_LE_UINT16( pPage->node.intNode.nKeys, pPage->iKeys );
    HB_PUT_LE_UINT32( pPage->node.intNode.leftPtr, pPage->Left );
    HB_PUT_LE_UINT32( pPage->node.intNode.rightPtr, pPage->Right );
